+++ src/coreComponents/physicsSolvers/fluidFlow/SinglePhaseStatistics.cpp
@@ -53,7 +53,7 @@ void SinglePhaseStatistics::registerDataOnMesh( Group & meshBodies )
   {
     ElementRegionManager & elemManager = mesh.getElemManager();
 
-    for( integer i = 0; i < regionNames.size(); ++i )
+    for( size_t i = 0; i < regionNames.size(); ++i )
     {
       ElementRegionBase & region = elemManager.getRegion( regionNames[i] );
       region.registerWrapper< RegionStatistics >( viewKeyStruct::regionStatisticsString() ).
@@ -99,7 +99,7 @@ void SinglePhaseStatistics::computeRegionStatistics( real64 const time,
 
   // Step 1: initialize the average/min/max quantities
   ElementRegionManager & elemManager = mesh.getElemManager();
-  for( integer i = 0; i < regionNames.size(); ++i )
+  for( size_t i = 0; i < regionNames.size(); ++i )
   {
     ElementRegionBase & region = elemManager.getRegion( regionNames[i] );
     RegionStatistics & regionStatistics = region.getReference< RegionStatistics >( viewKeyStruct::regionStatisticsString() );
@@ -213,7 +213,7 @@ void SinglePhaseStatistics::computeRegionStatistics( real64 const time,
   } );
 
   // Step 3: synchronize the results over the MPI ranks
-  for( integer i = 0; i < regionNames.size(); ++i )
+  for( size_t i = 0; i < regionNames.size(); ++i )
   {
     ElementRegionBase & region = elemManager.getRegion( regionNames[i] );
     RegionStatistics & regionStatistics = region.getReference< RegionStatistics >( viewKeyStruct::regionStatisticsString() );
